refactoring + add unit tests
